build: NetBSD fixes for use of the macro toupper() and tolower()
gcc on NetBSD fails when passing char* to the to{upper,lower} macros.
Cast to a int to work around this.

CR: dalmeida@likewise.com

(libkeytab: r39716)

+++ libkeytab/utils/ktstr.c
@@ -178,7 +178,7 @@ KtStrToUpper(
     }
 
     while (*pszString != '\0') {
-        *pszString = toupper(*pszString);
+        *pszString = toupper((int)*pszString);
         pszString++;
     }
 }
@@ -194,7 +194,7 @@ KtStrToLower(
     }
 
     while (*pszString != '\0') {
-        *pszString = tolower(*pszString);
+        *pszString = tolower((int)*pszString);
         pszString++;
     }
 }
